Add a systems metadata field
This field is used to described the system(s) that the document applies
to; there's an affordance for multiple systems in cases where a document
applies to more than one.

The system tags themselves are a controlled vocabulary described in
schema.systems.json. The systems are described as arrays of component
names; the system descriptors have increasing hierarchical specificity.

This helps to avoid confusion between what a "motor", for example,
belongs to.

Second, the format also makes easier to automatically group documents by
their parent systems.